Break up decl and impl in GCE util files.
This change is needed as part of the GCE credentials being added in one
of the next PRs in this chain; after putting those changes together,
this one was large and modular enough to split into its own PR.
