+++ google/cloud/storage/CMakeLists.txt
@@ -123,6 +123,7 @@ add_library(storage_client
             internal/complex_option.h
             internal/common_metadata.h
             internal/compute_engine_util.h
+            internal/compute_engine_util.cc
             internal/curl_handle.h
             internal/curl_handle.cc
             internal/curl_handle_factory.h
